+++ Runtime/SingletonLifetime.cs
@@ -1,0 +1,8 @@
+namespace Jeomseon.Unity.Singleton
+{
+    public enum SingletonLifetime
+    {
+        Application,
+        Scene
+    }
+}